pyproject.toml
[tool.poetry]
name = "heroes-4-1"
version = "0.1.0"
description = ""
authors = ["DanStaf <danstaf@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
psycopg2 = "^2.9.9"
pillow = "^10.4.0"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
aiogram = "^3.13.0"
django-apscheduler = "^0.6.2"
telebot = "^0.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
